requirements.txt
kafka-python==2.0.2
requests==2.31.0
cassandra-driver==3.28.0
pyspark==3.5.1
